Ship upstream keystone-paste.ini to /etc/keystone
It is the upstream expectation that keystone-paste.ini is a
configuration file that can be modified.

An example of this would be when using the admin_auth_token filter
to bootstrap keystone. The admin_auth_token filter should not be
in the pipeline beyond the initial bootstrapping and should be
removed afterwards. An example implementation of this can be
seen here [1].

Related bugs:
- https://bugs.launchpad.net/keystone/+bug/1545761
- https://bugs.launchpad.net/keystone/+bug/1545789

[1]: https://review.openstack.org/#/c/256248

Change-Id: I86b255cf2fd651db9c52954090c2d5dc6c85236e