Seperate entrypoint for tempest plugin for python-keystone-tests
Previously python-keystone package installed the entry point for the in-tree
Tempest plugin which cause failure while running tempest list-plugins
because Tempest found the entry point but does not have Tempest plugin
code which is subpackaged in python-keystone-tests.
Separate keystone_tests egg-info is created with the plugin entry point
so that Tempest includes the plugin only when subpackage is installed.

Change-Id: I19abfae297cbb09c015f45c0627afd8a410810b6