Ship an empty /etc/keystone/policy.json
Keystone project stopped to ship policy.json file, mainly because they
manage their policies in the code itself:
https://github.com/openstack/keystone/tree/master/keystone/common/policies

So they don't need a file to create policies.
In TripleO though, we want our operators able to change the policies
when needed, so we need that file to be shipped otherwise Puppet won't
be able to manage the policies.

This patch create an empty hash in policy.json and ship it as a
configuration file. Because it's empty, the default policies are not
overriden so there is no risk or backward compatibility issue in this
patch.

Change-Id: Ibd38e39ba133a2a7f321adc57f6d313f936202a5