Update patch set 1
Patch Set 1:

tl;dr we don't want to manage files in puppet because if we remove puppet we loose the files management. The place is in packaging. We had this discussion back in times when folks wanted to manage file permissions in puppet. we said no, and fix packaging.

Patch-set: 1