Generate and ship configuration schema in YAML + JSON
As part of the efforts documented in:
http://specs.openstack.org/openstack/oslo-specs/specs/pike/machine-readable-sample-config.html

We would like to provide to the operators / installers a Human Readable
version of services schemas so they can use the data in automation tools
if needed.

This patch will use oslo-config-generator to generate YAML & JSON
versions of configuration schemas and ship the 2 files in /user/share/<service>/<service>-schema.(json,yaml)

That way, operators and installers don't have to generate these file
themselves and they can directly consume it.

Change-Id: I68780639d7a254046589875337e84f3d5253b805