Update patch set 5
Patch Set 5:

IIUC, these new json and yaml files will be only used by tools to generate actual config files used by services, is that correct? or will these files be used as the actual config files by services?

Note that files in /etc/keystone/ should be only used for config files that could be editable for users. If this is not the case for these files and they are only a way to automatically document configuration options in a machine readable format, i think they should go to a different place, maybe as documentation files and they should not be marked as '%config(noreplace)'

Patch-set: 5
Label: Code-Review=0