Update patch set 1
Patch Set 1:

(1 comment)

2019-08-13 21:47:37.519137 | rdo-centos-7 | 2019-08-13 21:47:37,518 INFO:dlrn-build-mock:DEBUG: BUILDSTDERR: Error: Package: 1:python-keystone-14.1.1-0.20190813214154.c220cc4.el7.noarch (/python-keystone-14.1.1-0.20190813214154.c220cc4.el7.noarch)
2019-08-13 21:47:37.519289 | rdo-centos-7 | 2019-08-13 21:47:37,518 INFO:dlrn-build-mock:DEBUG: BUILDSTDERR:            Requires: python2-bcrypt >= 3.1.3
2019-08-13 21:47:37.519452 | rdo-centos-7 | 2019-08-13 21:47:37,518 INFO:dlrn-build-mock:DEBUG: BUILDSTDERR:            Available: python2-bcrypt-3.1.2-3.el7.x86_64 (delorean-rocky-testing)
2019-08-13 21:47:37.519579 | rdo-centos-7 | 2019-08-13 21:47:37,518 INFO:dlrn-build-mock:DEBUG: BUILDSTDERR:                python2-bcrypt = 3.1.2-3.el7

bcrypt in rocky is 3.1.2 currently so DLRN is failing. Is the bcrypt bump to 3.1.3 really necessary, 3.1.2 is not sufficient, if not can you try without it?

Looks like you need this for some bug, can u also share that bug. Asking to understand the need so late in cycle and change might also be needed changes in other releases as well.

Patch-set: 1
Label: Workflow=0